Generating random index to select a random definition.
For testing, I used this random input to populate the buttons with it's respective answer.

At the end of the function, I remove the id from the List, when the list is empty, Intent to the score screen. (This part is being tricky so it'd probably be 10x easier to just post the score on the Quiz page)