Finished update to transfer this to being a quiz on Hockey player salaries. A few bugs from the old Quiz were fixed.
Soft shuffling was implemented to reduce the chance of "C" and "D" buttons not being set on creating. Soft shuffling shuffles the values 10 times before the quiz starts and doesn't remove the definitions so that instead of a 1/3 chance of the buttons not being set it's more like a 1/1000 chance.